requirements.cloud.txt
fastapi
uvicorn[standard]

langchain
langchain-community
langchain-core
langchain-huggingface
langchain_chroma
langchain-text-splitters

chromadb

sentence-transformers

pypdf
python-dotenv
python-multipart

openai